+++ ivmodels/models/__init__.py
@@ -1,11 +1,9 @@
 from .anchor_regression import AnchorRegression
 from .kclass import KClass
 from .pulse import PULSE
-from .space_iv import SpaceIV
 
 __all__ = [
     "AnchorRegression",
     "KClass",
     "PULSE",
-    "SpaceIV",
 ]